Merge staging into main

+++ api/index.py
@@ -20,10 +20,10 @@ chatgpt = ChatGPT()
 
 # region Language related
 
-translate_language = "Japanese"
-audio_language = "Traditional Chinese"
+default_translate_language = "Japanese"
+default_audio_language = "Traditional Chinese"
 
-lan_dic = {
+lang_dict = {
     "繁體中文": "Traditional Chinese",
     "英文": "English",
     "日文": "Japanese",
@@ -35,7 +35,16 @@ lan_dic = {
     "荷蘭文": "Dutch",
     "德文": "German"
 }
-reverse_lan_dict = {value: key for key, value in lan_dic.items()}
+reverse_lang_dict = {value: key for key, value in lang_dict.items()}
+
+# endregion
+
+# region User related
+
+user_translate_language_key = 'translate_language'
+user_audio_language_key = 'audio_language'
+
+user_dict = {}
 
 # endregion
 
@@ -59,9 +68,13 @@ def callback():
 
 @line_handler.add(MessageEvent, message=TextMessage)
 def handle_text_message(event):
-    global translate_language, audio_language
     user_input = event.message.text
     if (user_input == "/setting") or (user_input == "設定"):
+        # Set default language by user
+        user_dict[event.source.user_id] = {
+            user_translate_language_key: default_translate_language,
+            user_audio_language_key: default_audio_language
+        }
         flex_message = TextSendMessage(text="請選擇語音辨識後的翻譯語言（我方語言）",
                                        quick_reply=QuickReply(items=[
                                            QuickReplyButton(action=MessageAction(
@@ -88,7 +101,9 @@ def handle_text_message(event):
         line_bot_api.reply_message(event.reply_token, flex_message)
 
     elif ("設定辨識翻譯" in user_input):
-        audio_language = lan_dic[user_input.split(" ")[1]]
+        # Set audio language by user
+        user_dict[event.source.user_id][user_audio_language_key] = lang_dict[user_input.split(" ")[
+            1]]
         flex_message = TextSendMessage(text="請選擇打字後的翻譯語言（對方語言）",
                                        quick_reply=QuickReply(items=[
                                            QuickReplyButton(action=MessageAction(
@@ -115,40 +130,69 @@ def handle_text_message(event):
         line_bot_api.reply_message(event.reply_token, flex_message)
 
     elif ("設定打字翻譯" in user_input):
-        translate_language = lan_dic[user_input.split(" ")[1]]
+        user_id = event.source.user_id
+        # Set translate language by user
+        user_dict[user_id][user_translate_language_key] = lang_dict[user_input.split(" ")[
+            1]]
+        # Format response message
+        audio_language = user_dict[user_id][user_audio_language_key]
+        translate_language = user_dict[user_id][user_translate_language_key]
         response = f"""設定完畢！
-我方語言：{reverse_lan_dict[audio_language]}（{audio_language}）
-對方語言：{reverse_lan_dict[translate_language]}（{translate_language}）"""
+我方語言：{reverse_lang_dict[audio_language]}（{audio_language}）
+對方語言：{reverse_lang_dict[translate_language]}（{translate_language}）"""
         line_bot_api.reply_message(
             event.reply_token, TextSendMessage(text=response))
 
     elif (user_input == "/current-setting") or (user_input == "目前設定"):
-        response = f"""我方語言：{reverse_lan_dict[audio_language]}（{audio_language}）
-對方語言：{reverse_lan_dict[translate_language]}（{translate_language}）"""
-        line_bot_api.reply_message(
-            event.reply_token, TextSendMessage(text=response))
+        user_id = event.source.user_id
+        if not (user_exists(user_id)):
+            line_bot_api.reply_message(
+                event.reply_token, TextSendMessage(text="請先輸入「/setting」或「設定」指令，設定您的翻譯情境！"))
+        else:
+            # Format response message
+            audio_language = user_dict[user_id][user_audio_language_key]
+            translate_language = user_dict[user_id][user_translate_language_key]
+            response = f"""我方語言：{reverse_lang_dict[audio_language]}（{audio_language}）
+對方語言：{reverse_lang_dict[translate_language]}（{translate_language}）"""
+            line_bot_api.reply_message(
+                event.reply_token, TextSendMessage(text=response))
 
     else:
-        response = chatgpt.translate(user_input, translate_language)
-        line_bot_api.reply_message(
-            event.reply_token, TextSendMessage(text=response))
+        user_id = event.source.user_id
+        if not (user_exists(user_id)):
+            line_bot_api.reply_message(
+                event.reply_token, TextSendMessage(text="請先輸入「/setting」或「設定」指令，設定您的翻譯情境！"))
+        else:
+            response = chatgpt.translate(
+                user_input, user_dict[user_id][user_translate_language_key])
+            line_bot_api.reply_message(
+                event.reply_token, TextSendMessage(text=response))
 
 
 @line_handler.add(MessageEvent, message=AudioMessage)
 def handle_audio_message(event):
-    global translate_language, audio_language
-    message_id = event.message.id
-    message_content = line_bot_api.get_message_content(message_id)
-    user_audio_path = os.path.join(app.config.get(
-        'AUDIO_BASE_PATH'), f'{message_id}.m4a')
-    with open(user_audio_path, 'wb') as f:
-        f.write(message_content.content)
-    whisper_text = chatgpt.whisper(user_audio_path)
-    if (os.path.exists(user_audio_path)):
-        os.remove(user_audio_path)
-    response_text = chatgpt.translate(whisper_text, audio_language)
-    line_bot_api.reply_message(
-        event.reply_token, TextSendMessage(text=response_text))
+    user_id = event.source.user_id
+    if not (user_exists(user_id)):
+        line_bot_api.reply_message(
+            event.reply_token, TextSendMessage(text="請先輸入「/setting」或「設定」指令，設定您的翻譯情境！"))
+    else:
+        message_id = event.message.id
+        message_content = line_bot_api.get_message_content(message_id)
+        user_audio_path = os.path.join(app.config.get(
+            'AUDIO_BASE_PATH'), f'{message_id}.m4a')
+        with open(user_audio_path, 'wb') as f:
+            f.write(message_content.content)
+        whisper_text = chatgpt.whisper(user_audio_path)
+        if (os.path.exists(user_audio_path)):
+            os.remove(user_audio_path)
+        response_text = chatgpt.translate(
+            whisper_text, user_dict[user_id][user_audio_language_key])
+        line_bot_api.reply_message(
+            event.reply_token, TextSendMessage(text=response_text))
+
+
+def user_exists(user_id):
+    return user_id in user_dict
 
 
 if __name__ == '__main__':
